// src/actions/types/tables.ts
export enum types {
  OPEN_EDIT_MODAL = "OPEN_EDIT_MODAL",
  CLOSE_EDIT_MODAL = "CLOSE_EDIT_MODAL",
  RESET_EDIT_FORM = "RESET_EDIT_FORM",
  ADD_TABLE_RECORD = "ADD_TABLE_RECORD",
  EDIT_TABLE_RECORD = "EDIT_TABLE_RECORD",
  DELETE_TABLE_RECORD = "DELETE_TABLE_RECORD",
  EDIT_TABLE_CLONE_RECORD = "EDIT_TABLE_CLONE_RECORD",
  DELETE_TABLE_CLONE_RECORD = "DELETE_TABLE_CLONE_RECORD",
  COPY_TABLE = "COPY_TABLE",
  DELETE_TABLE = "DELETE_TABLE",
}